apps/agents/requirements.txt
google-cloud-aiplatform>=1.60.0
google-adk>=0.3.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0
google-cloud-storage>=2.17.0
google-cloud-bigquery>=3.19.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-gcp-trace>=1.6.0
pydantic>=2.7.0
fastapi>=0.111.0
uvicorn>=0.29.0
